storing to revert

+++ src/main/java/com/techelevator/model/RestaurantDao.java
@@ -14,9 +14,9 @@ public interface RestaurantDao {
 
     public void addEventToTable(List<Long> restaurantId, String username, LocalDate deadline);
 
-    public List<Restaurant> getRestaurantsByEventId(int eventId);
+    public List<Restaurant> getRestaurantsByEventId(Long eventId);
 
-    public boolean isWithinDeadline(int eventId);
+    public boolean isWithinDeadline(Long eventId);
 
 
 }
